backend/ai-engine/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
openai==1.3.8
langchain==0.0.350
langchain-openai==0.0.2
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
scipy==1.11.4
plotly==5.17.0
redis==5.0.1
celery==5.3.4
sqlalchemy==2.0.23
asyncpg==0.29.0
httpx==0.25.2
websockets==12.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-dotenv==1.0.0
aiofiles==23.2.1
pytz==2023.3
python-dateutil==2.8.2

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2
factory-boy==3.3.0
faker==20.1.0
responses==0.24.1